+++ quizs/src/main/java/com/quizly/quizs/repositories/QuizRepository.java
@@ -1,6 +1,5 @@
 package com.quizly.quizs.repositories;
 
-import com.quizly.quizs.dtos.QuizDto;
 import com.quizly.quizs.models.Quiz;
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -10,5 +9,5 @@ import java.util.Optional;
 public interface QuizRepository extends JpaRepository<Quiz,Long> {
     Optional<Quiz> findById(Long id);
     List<Quiz> findQuizzesByOwner(String owner);
-    int countQuizzesByOwner(String owner);
+    Long countQuizzesByOwner(String owner);
 }
